+++ src/main/kotlin/racingcar/domain/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package model
+package racingcar.domain
 
 interface NumberGenerator {
     fun generate(): Int
