+++ src/main/kotlin/racingcar/domain/model/AttemptCount.kt
@@ -1,4 +1,4 @@
-package model
+package racingcar.domain.model
 
 class AttemptCount(private val input: String) {
     init {
