+++ src/main/kotlin/racingcar/domain/model/Car.kt
@@ -1,4 +1,4 @@
-package model
+package racingcar.domain.model
 
 class Car(
     val name: String,
