+++ src/test/kotlin/racingcar/domain/model/AttemptCountTest.kt
@@ -1,4 +1,4 @@
-package model
+package racingcar.domain.model
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
