+++ src/test/kotlin/racingcar/domain/model/CarsTest.kt
@@ -1,10 +1,11 @@
-package model
+package racingcar.domain.model
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
 import org.junit.jupiter.api.assertThrows
+import racingcar.domain.NumberGenerator
 
 class CarsTest {
     @Test
